Implement CSVTuple and make HeapTuple interface.
Currently the catalog data is CSV.  It is desirable to make all low-layer
access interface and pluggable, so that sometime (such like in test) CSV
can be one option.
Add a simple test, too.  The go test command sits in the package directory
when executing, so the test data resides below access.